+++ web-console/src/app/(root)/(blank)/login/page.tsx
@@ -1,0 +1,18 @@
+'use client'
+
+import { redirect } from 'next/navigation'
+
+import { Authenticator } from '@aws-amplify/ui-react'
+import { Box } from '@mui/material'
+
+export default () => {
+  return (
+    <Box sx={{ display: 'flex', height: '100vh', justifyContent: 'center' }}>
+      <Authenticator>
+        {({}) => {
+          redirect('/home')
+        }}
+      </Authenticator>
+    </Box>
+  )
+}